# Speed up start up and reload times