@jvican 2017-04-18
The error https://travis-ci.org/sbt/sbt/jobs/223197672 is reproducible from latest master locally, which hints this PR is not causing it.