@jvican 2017-04-18
The error has been fixed and a more stable implementation of the `autoImport` detector has been added.
  
I think GitHub has a bug in the UI and is showing the commits in an incorrect order. The CI has kicked in for the last one.