@jvican 2017-04-18
The correct order is here. https://github.com/scalacenter/sbt/commits/speed-up-startup-time-1.0