@jvican 2017-04-18
@eed3si9n @dwijnand This is ready for review again. Please, can you have a look?